refactor: modularize save utilities
